chore(bdev): reorder the location of some files
The bdev dir contains all bdevs we use. This gets rid of the extra
dev directory which just made things more confusing.
